requirements.txt
tensorflow==2.4.1
keras==2.3.1
numpy==1.19.5
ray==2.6.1
click==8.1.6
natsort==8.4.0
monai[itk,nibabel,matplotlib]==1.1.0
Pillow==10.0.0
scikit-image==0.19.3
wcmatch==8.4.1
SimpleITK==2.2.1